+++ cub/test/catch2_test_device_transform.cu
@@ -331,7 +331,7 @@ struct sum_five
   __host__ __device__ auto operator()(std::int8_t a, std::int16_t b, std::int32_t c, std::int64_t d, float e) const
     -> double
   {
-    return a + b + c + d + e;
+    return a + b + c + static_cast<float>(d) + e;
   }
 };
